bot.give: Using cache counter is not correct
When the cache-counter is used, user can give any huge number of nutshells
in a very short of time. This is because the time only matters after the
maximum number of counter is reached; it has no effect if the counter is
still small. The correct way is to have two variables in cache database.